在clang++下通过编译和测试

+++ Src/Math/Tri.h
@@ -23,7 +23,7 @@ template<typename T> AGZ_FORCE_INLINE auto AsRad(Deg<T> deg) { return Deg2Rad<T>
 template<typename T> AGZ_FORCE_INLINE auto AsRad(Rad<T> rad) { return rad; }
 
 template<typename T> AGZ_FORCE_INLINE T Sin_rawimpl(T);
-template<>           AGZ_FORCE_INLINE float  Sin_rawimpl<float>(float rad) { return std::sinf(rad); }
+template<>           AGZ_FORCE_INLINE float  Sin_rawimpl<float>(float rad) { return std::sin(rad); }
 template<>           AGZ_FORCE_INLINE double Sin_rawimpl<double>(double rad) { return std::sin(rad); }
 
 template<typename T> struct Sin_impl;
@@ -33,7 +33,7 @@ template<typename T> struct Sin_impl<Rad<T>> { static AGZ_FORCE_INLINE auto Sin(
 template<typename T> struct Sin_impl<Deg<T>> { static AGZ_FORCE_INLINE auto Sin(Deg<T> deg) { return Sin_rawimpl<T>(deg.value * (PI_impl<T>::PI() / T(180.0))); } };
 
 template<typename T> AGZ_FORCE_INLINE T Cos_rawimpl(T);
-template<>           AGZ_FORCE_INLINE float  Cos_rawimpl<float>(float rad) { return std::cosf(rad); }
+template<>           AGZ_FORCE_INLINE float  Cos_rawimpl<float>(float rad) { return std::cos(rad); }
 template<>           AGZ_FORCE_INLINE double Cos_rawimpl<double>(double rad) { return std::cos(rad); }
 
 template<typename T> struct Cos_impl;
